Revert "i386/amd64: fix LP64 format strings, pointer types, and paddr_t in regression test"
This reverts commit e9b1060fa72cbdad783c39e05f32e9b7643d42b2.

+++ minix/commands/rawspeed/rawspeed.c
@@ -220,8 +220,8 @@ int main(int argc, char **argv)
 	if (nbytes > 0) {
 		off_t kBpts;
 
-		fprintf(stderr, "%lld kB / %ld.%ld s = ",
-			(long long)(nbytes + 512) / 1024,
+		fprintf(stderr, "%ld kB / %ld.%ld s = ",
+			(nbytes + 512) / 1024,
 			tenthsec / 10, tenthsec % 10);
 		if (tenthsec < 5)
 			fprintf(stderr, "infinite\n");
@@ -230,12 +230,12 @@ int main(int argc, char **argv)
 				seconds = (tenthsec + 5) / 10;
 				kBpts= (nbytes + 512L * seconds)
 							/ (1024L * seconds);
-				fprintf(stderr, "%lld kB/s\n", (long long)kBpts);
+				fprintf(stderr, "%ld kB/s\n", kBpts);
 			} else {
 				kBpts= (100 * nbytes + 512L * tenthsec)
 							/ (1024L * tenthsec);
-				fprintf(stderr, "%lld.%d kB/s\n",
-					(long long)(kBpts/10), (int)(kBpts%10));
+				fprintf(stderr, "%ld.%d kB/s\n",
+					kBpts/10, (int)(kBpts%10));
 			}
 		}
 	}
@@ -246,10 +246,10 @@ int main(int argc, char **argv)
 		tenthms= (tenthsec * 1000 + nseeks/2) / nseeks;
 
 		fprintf(stderr,
-		"%lld seeks / %ld.%ld s = %lld seeks/s = %lld.%d ms/seek\n",
-			(long long)nseeks, tenthsec / 10, tenthsec % 10,
-			(long long)(nseeks * 10 + tenthsec/2) / tenthsec,
-			(long long)(tenthms / 10), (int)(tenthms % 10));
+		"%ld seeks / %ld.%ld s = %ld seeks/s = %ld.%d ms/seek\n",
+			nseeks, tenthsec / 10, tenthsec % 10,
+			(nseeks * 10 + tenthsec/2) / tenthsec,
+			tenthms / 10, (int)(tenthms % 10));
 
 		for (rpm= 3600; rpm <= 7200; rpm+= 1800) {
 			int rotms = (10000L / 2 * 60 + rpm/2) / rpm;
@@ -263,8 +263,8 @@ int main(int argc, char **argv)
 			} else {
 				fprintf(stderr, ", ");
 			}
-			fprintf(stderr, "%lld.%d ms (%d rpm)",
-				(long long)(tenthms - rotms) / 10,
+			fprintf(stderr, "%ld.%d ms (%d rpm)",
+				(tenthms - rotms) / 10,
 				(int)((tenthms - rotms) % 10),
 				rpm);
 		}
